BUG#: 3211
TITLE:Client Performance Statistics Test Fails

DESCRIPTION: checking in code to fix test case and checking in code to move statisitcs gathering from CIMDateTime based to TimeValue based

+++ src/Pegasus/Common/XmlWriter.h
@@ -270,7 +270,7 @@ public:
         HttpMethod httpMethod,
     const ContentLanguages & contentLanguages,
         Uint32 contentLength,
-	Uint64 serverResponseTime = 0);
+	Uint32 serverResponseTime = 0);
 
     static void appendHttpErrorResponseHeader(
 	Array<char>& out,
@@ -379,7 +379,7 @@ public:
         HttpMethod httpMethod,
         const ContentLanguages & httpContentLanguages,          
 	const Array<char>& body,
-	Uint64 serverResponseTime,
+	Uint32 serverResponseTime,
 	Boolean isFirst = true,
 	Boolean isLast = true);
 
@@ -407,7 +407,7 @@ public:
         HttpMethod httpMethod,
         const ContentLanguages & httpContentLanguages,  
 	const Array<char>& body,
-	Uint64 serverResponseTime,
+	Uint32 serverResponseTime,
 	Boolean isFirst = true,
 	Boolean isLast = true);
